Zoom out if necessary to show a searched comparison point.
Not doing this for a reference point search because we either center map on
found location or sync map with another. So you can see maps are updated in
response to your action. But for a comparison point outside of the map we need
to show somehow we've placed a marker there.